@@ -20,36 +20,37 @@
     "@fullcalendar/list": "^6.1.17",
     "@fullcalendar/timegrid": "^6.1.17",
     "@fullcalendar/vue3": "^6.1.17",
+    "@nuxt/devtools": "^2.4.0",
     "@nuxt/image": "^1.10.0",
     "@nuxt/ui": "^3.1.0",
     "@nuxtjs/color-mode": "^3.5.2",
     "@nuxtjs/i18n": "^9.5.3",
-    "@tailwindcss/postcss": "^4.1.4",
-    "@vueuse/core": "^10.7.0",
-    "firebase": "^10.7.0",
+    "firebase": "^10.14.1",
     "nuxt": "^3.16.2",
     "pinia": "^2.1.7",
     "sass": "^1.87.0",
-    "vue": "^3.3.8",
-    "vue-router": "^4.2.5"
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.1"
   },
   "devDependencies": {
-    "@nuxtjs/tailwindcss": "^6.13.2",
     "@rushstack/eslint-patch": "^1.3.3",
-    "@types/node": "^20.10.0",
-    "@vitejs/plugin-vue": "^5.0.0",
-    "@volar/typescript": "^1.8.22",
+    "@types/node": "^22.15.2",
+    "@vitejs/plugin-vue": "^5.2.3",
+    "@volar/typescript": "^2.4.13",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/eslint-config-typescript": "^12.0.0",
-    "@vue/tsconfig": "^0.4.0",
+    "@vue/tsconfig": "^0.7.0",
+    "@vueuse/core": "^13.1.0",
     "autoprefixer": "^10.4.21",
     "cssnano": "^7.0.6",
-    "eslint": "^8.49.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint": "^8.57.1",
+    "eslint-plugin-vue": "^9.33.0",
+    "postcss": "^8.5.3",
     "postcss-import": "^16.1.0",
-    "prettier": "^3.0.3",
+    "prettier": "^3.5.3",
+    "tailwindcss": "^3.4.1",
     "typescript": "^5.8.3",
     "vite": "^6.3.3",
-    "vue-tsc": "^1.8.22"
+    "vue-tsc": "^2.2.10"
   }
 }
