@@ -1,66 +1,15 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
-    './components/**/*.{js,vue,ts}',
-    './layouts/**/*.vue',
-    './pages/**/*.vue',
-    './plugins/**/*.{js,ts}',
-    './app.vue',
-    './error.vue'
+    "./components/**/*.{js,vue,ts}",
+    "./layouts/**/*.vue",
+    "./pages/**/*.vue",
+    "./plugins/**/*.{js,ts}",
+    "./app.vue",
+    "./error.vue"
   ],
   theme: {
-    extend: {
-      colors: {
-        primary: {
-          50: 'var(--primary-50)',
-          100: 'var(--primary-100)',
-          200: 'var(--primary-200)',
-          300: 'var(--primary-300)',
-          400: 'var(--primary-400)',
-          500: 'var(--primary-500)',
-          600: 'var(--primary-600)',
-          700: 'var(--primary-700)',
-          800: 'var(--primary-800)',
-          900: 'var(--primary-900)',
-        },
-        secondary: {
-          50: 'var(--secondary-50)',
-          100: 'var(--secondary-100)',
-          200: 'var(--secondary-200)',
-          300: 'var(--secondary-300)',
-          400: 'var(--secondary-400)',
-          500: 'var(--secondary-500)',
-          600: 'var(--secondary-600)',
-          700: 'var(--secondary-700)',
-          800: 'var(--secondary-800)',
-          900: 'var(--secondary-900)',
-        },
-        accent: {
-          DEFAULT: 'var(--accent)',
-          dark: 'var(--accent-dark)',
-        },
-        danger: {
-          DEFAULT: 'var(--danger)',
-          dark: 'var(--danger-dark)',
-        },
-        success: {
-          DEFAULT: 'var(--success)',
-          dark: 'var(--success-dark)',
-        },
-        warning: {
-          DEFAULT: 'var(--warning)',
-          dark: 'var(--warning-dark)',
-        },
-        info: {
-          DEFAULT: 'var(--info)',
-          dark: 'var(--info-dark)',
-        },
-      },
-      fontFamily: {
-        sans: ['Inter var', 'sans-serif']
-      },
-    },
+    extend: {},
   },
   plugins: [],
-  darkMode: 'class',
 } 
