@@ -1,8 +1,7 @@
 // Generated by nuxi
-/// <reference types="@nuxtjs/color-mode" />
 /// <reference types="@nuxt/ui" />
 /// <reference types="@nuxt/image" />
-/// <reference types="@nuxtjs/i18n" />
+/// <reference types="@nuxt/devtools" />
 /// <reference types="@nuxt/devtools" />
 /// <reference types="@nuxt/telemetry" />
 /// <reference path="types/builder-env.d.ts" />
