@@ -1,79 +1,48 @@
-// https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
+  compatibilityDate: "2024-11-01",
   devtools: { enabled: true },
-
-  modules: [
-    '@nuxtjs/postcss',
-    '@nuxtjs/color-mode',
-    '@nuxt/ui',
-    '@nuxt/image',
-    '@nuxtjs/i18n'
-  ],
-
-
-
-  colorMode: {
-    preference: 'system',
-    fallback: 'light',
-    classSuffix: ''
-  },
-
-  postcss: {
-    plugins: {
-      options.postcss
-    },
-  },
-
-
-  i18n: {
-    baseUrl: 'http://localhost:3000',
-    defaultLocale: 'en',
-    strategy: 'no_prefix',
-    detectBrowserLanguage: false,
-    locales: [
-      {
-        code: 'en',
-        name: 'English',
-        iso: 'en-US'
-      }
-    ],
-    vueI18n: './i18n.config.ts'
-  },
-
   vite: {
+    plugins: [
+    ],
     define: {
       'process.env.DEBUG': false,
     },
-    css: {
-      preprocessorOptions: {
-        scss: {
-          additionalData: '@use "@/assets/scss/_variables.scss" as *;'
-        }
-      }
-    },
-    optimizeDeps: {
-      include: [
-        '@fullcalendar/core',
-        '@fullcalendar/vue3',
-        '@fullcalendar/daygrid',
-        '@fullcalendar/interaction',
-        '@fullcalendar/timegrid',
-        '@fullcalendar/list'
-      ]
-    },
-    ssr: {
-      noExternal: ['@fullcalendar/core']
-    }
   },
-
+  css: ['~/assets/css/main.css'],
+  build: {
+    transpile: ['@fullcalendar/core']
+  },
+  modules: [
+    ['@nuxtjs/color-mode', {
+      preference: 'system',
+      fallback: 'light',
+      classSuffix: ''
+    }],
+    '@nuxt/ui',
+    '@nuxt/image',
+    ['@nuxtjs/i18n', {
+      baseUrl: 'http://localhost:3000',
+      defaultLocale: 'en',
+      strategy: 'no_prefix',
+      detectBrowserLanguage: false,
+      locales: [
+        {
+          code: 'en',
+          name: 'English',
+          iso: 'en-US'
+        }
+      ],
+      vueI18n: './i18n.config.ts'
+    }],
+    '@nuxt/devtools'
+  ],
   plugins: [
     { src: '~/plugins/fullcalendar.ts', mode: 'client' }
   ],
-
   typescript: {
     strict: true,
     typeCheck: true
-  },
-
-  compatibilityDate: '2025-04-25'
+  }
 })
+
+
