@@ -1,8 +1,6 @@
 module.exports = {
   plugins: {
-   
-    '@tailwindcss/postcss': {},
+    'tailwindcss': {},
     autoprefixer: {},
-    ...(process.env.NODE_ENV === 'development' ? { cssnano: { preset: 'default' } } : {})
   }
-} 
+}
